requirements.txt
beautifulsoup4==4.6.0
dj-database-url==0.5.0
Django==2.0.5
django-allauth==0.36.0
django-environ==0.4.5
django-heroku==0.3.1
django-multiselectfield==0.1.8
django-rest-auth==0.9.3
djangorestframework==3.7.7
future==0.16.0
geocoder==1.38.1
geographiclib==1.49
geopy==1.16.0
gunicorn==19.9.0
Pillow==4.2.1
psycopg2==2.7.5
requests==2.18.1
requests-oauthlib==1.0.0
virtualenv==16.0.0
whitenoise==3.3.1
